Update CFClient interface to reflect changes in Java server side SDK
