+++ client/api/Message.cs
@@ -1,0 +1,11 @@
+﻿using System;
+namespace io.harness.cfsdk.client.api
+{
+    public class Message
+    {
+        public string Event { get; set; }
+        public string Domain { get; set; }
+        public string Identifier { get; set; }
+        public int Version { get; set; }
+    }
+}
