+++ client/connector/Connector.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace io.harness.cfsdk.client.connector
+{
+    public interface IConnector
+    {
+
+    }
+}
